+++ src/services/import/single.js
@@ -3,7 +3,7 @@
 const noteService = require('../../services/notes');
 const imageService = require('../../services/image');
 const protectedSessionService = require('../protected_session');
-const markdownService = require('./markdown.js');
+const markdownService = require('./markdown');
 const mimeService = require('./mime');
 const utils = require('../../services/utils');
 const importUtils = require('./utils');
